requirements.txt
# Django==4.2.0
# djangorestframework==3.14.0
# psycopg2-binary==2.9.6
# python-decouple==3.8
# drf-spectacular==0.26.3
# django-cors-headers==4.0.0
Django==4.2.0
djangorestframework==3.14.0
python-decouple==3.8
drf-spectacular==0.26.3
django-cors-headers==4.0.0